Solution to Problem 11
This Problem is bit tricky . 
Prior to this code , I wrote a small code to get the array in the desired form . As in the question it's given with spaces(' ')and enter('\n'), so to get the in this form what I do is eliminate the '\n' than removed ' ' with the if statement. After that I copy and pasted the output in the declaration of the int array.
Also note that we have remove zero before a no. because in C language zero before a no. is considered to be in octal number system , So 09 is not considered as 9 instead it's shows some octal error that 09 doesn't exist in octal system . So , I manually removed the 0 before the number .